lcm: set PT level to current node level
When a clause is modified via LCM, tracking the PT level is difficult.
As an over approximation, use the level of the current node. This way, we
can at least use LCM, and still share clauses downwards correctly. For now,
this is good enough, compared to not being able to share clauses at all.